Move stackedContext out of the breadcrumbs subfolder
