+++ static/app/components/replays/stackedContent.tsx
